Implementation of memcpy and memmove function
The important difference is that it is undefined behavior to call memcpy with overlapping regions. 
One must use memmove for that. As the names imply, memcpy copies data from one region to another, 
while memmove moves data within a region. (It’s also perfectly acceptable to memmove between 
different regions.